Modified the target frameworks for the projects
This adds a .NET standard 2.1 target in place of the .NET core 3.1 target, and also includes a .NET framework 4.6.1 target on the test project.

The .NET 5 target was left in-place despite being out-of-support as that is the first .NET version providing the `System.Half` type, and this library provides simple low-level information about that type that may be useful to developers still stuck on .NET 5 for whatever reason.